add gradle dependcy: Glide, litesuites-common-utils
