+++ app/build.gradle
@@ -37,4 +37,11 @@ dependencies {
     compile 'com.android.support:appcompat-v7:23.4.0'
     compile 'com.android.support:support-v4:23.0.1'
     compile 'com.squareup.okhttp3:okhttp:3.2.0'
+    // image tool
+    compile 'com.github.bumptech.glide:glide:3.7.0'
+    // util
+    compile 'com.luffykou:android-common-utils:1.1.3'
+
+    debugCompile 'com.squareup.leakcanary:leakcanary-android:1.3.1'
+    releaseCompile 'com.squareup.leakcanary:leakcanary-android-no-op:1.3.1'
 }
