+++ app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.appbyme.app178470"
+    package="com.appbyme.dev"
     android:versionCode="10103"
     android:versionName="1.1.3">
