+++ app/src/main/java/com/mobcent/discuz/fragments/DiscoveryFragment.java
@@ -1,14 +1,11 @@
 package com.mobcent.discuz.fragments;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AbsListView;
-import android.widget.ListView;
 
-import com.appbyme.app178470.R;
+import com.appbyme.dev.R;
 
 /**
  * Created by ubuntu on 16-6-21.
