+++ app/src/main/java/com/mobcent/discuz/fragments/DiscuzFragment.java
@@ -1,7 +1,6 @@
 package com.mobcent.discuz.fragments;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.support.v4.view.PagerAdapter;
 import android.support.v4.view.PagerTabStrip;
 import android.support.v4.view.ViewPager;
@@ -9,10 +8,8 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AbsListView;
-import android.widget.ListView;
 
-import com.appbyme.app178470.R;
+import com.appbyme.dev.R;
 
 import java.util.ArrayList;
