+++ app/src/main/java/com/mobcent/discuz/fragments/ZhidaoFragment.java
@@ -7,7 +7,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.appbyme.app178470.R;
+import com.appbyme.dev.R;
 
 /**
  * 知道碎片界面
