+++ app/src/test/java/com/appbyme/dev/ExampleUnitTest.java
@@ -1,0 +1,15 @@
+package com.appbyme.dev;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * To work on unit tests, switch the Test Artifact in the Build Variants view.
+ */
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() throws Exception {
+        assertEquals(4, 2 + 2);
+    }
+}
